+++ requirements.txt
@@ -1,6 +1,17 @@
 asn1crypto==0.24.0
+astroid==2.2.5
 cffi==1.12.3
 cryptography==2.7
+isort==4.3.21
+lazy-object-proxy==1.4.2
+mccabe==0.6.1
+mypy==0.720
+mypy-extensions==0.4.1
+pep8==1.7.1
 pycparser==2.19
+pylint==2.3.1
 six==1.12.0
 stdiomask==0.0.3
+typed-ast==1.4.0
+typing-extensions==3.7.4
+wrapt==1.11.2
